feat: Make tests work without migrations

+++ src/__tests__/Transaction.spec.ts
@@ -13,12 +13,14 @@ let connection: Connection;
 describe('Transaction', () => {
   beforeAll(async () => {
     connection = await createConnection('test-connection');
-    
-    await connection.query('DROP TABLE IF EXISTS transactions');
-    await connection.query('DROP TABLE IF EXISTS categories');
-    await connection.query('DROP TABLE IF EXISTS migrations');
-    
-    await connection.runMigrations();
+
+    if (connection.options.synchronize !== true) {
+      await connection.query('DROP TABLE IF EXISTS transactions');
+      await connection.query('DROP TABLE IF EXISTS categories');
+      await connection.query('DROP TABLE IF EXISTS migrations');
+
+      await connection.runMigrations();
+    }
   });
 
   beforeEach(async () => {
